refactor: modularize friend requests into FriendRequestsList component
