+++ src/app/(protected)/friends/requests.tsx
@@ -1,46 +1,16 @@
-import { useRouter, Stack } from "expo-router";
+import { Stack } from "expo-router";
 import { useEffect, useState, useRef } from 'react';
-import { View, FlatList, Text, TouchableOpacity, ActivityIndicator, Animated, Image } from "react-native";
+import { View, Animated } from "react-native";
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { FriendRequest } from "@/types";
 import FriendRequestModal from "@/components/friendRequestModal";
+import FriendRequestsList from "@/components/friendRequestsList";
   
 
 export default function RequestsScreen() {
-    const router = useRouter();
-    const [requests, setRequests] = useState<FriendRequest[]>([]);
-    const [loading, setLoading] = useState(false);
-
     const [selectedRequest, setSelectedRequest] = useState<FriendRequest | null>(null);
     const [modalVisible, setModalVisible] = useState(false);
     const scaleAnim = useRef(new Animated.Value(0)).current;
-
-    const listFriendRequests = async () => {
-        try {
-            setLoading(true);
-            const token = await AsyncStorage.getItem("access_token");
-            const res = await fetch(`${process.env.EXPO_PUBLIC_API_URL}/friends/requests`, {
-                method: "GET",
-                headers: {
-                    "Content-Type": "application/json",
-                    Authorization: `Bearer ${token}`,
-                }
-            });
-
-            if (!res.ok) {
-                const { error } = await res.json();
-                console.error("Error fetching requests:", error);
-                return;
-            }
-
-            const data = await res.json();
-            setRequests(data);
-        } catch(error) {
-            console.error('Requests error:', 'failed to retrieve requests');
-        } finally {
-            setLoading(false);
-        }
-    }
 
     const openModal = (request: FriendRequest) => {
         setSelectedRequest(request);
@@ -87,77 +57,35 @@ export default function RequestsScreen() {
                 console.error("Error fetching requests:", error);
                 return;
             }
-
-            const data = await res.json();
-            listFriendRequests();
         } catch (error) {
             console.error('Error', error);
         }
     }
-    
-    useEffect(() => {
-        listFriendRequests();
-    }, [])
 
     return (
         <>
         <Stack.Screen/>
         
         <View className="flex-1 bg-black px-4 pt-6">
-            {loading ? (
-                <View className="flex-1 justify-center items-center">
-                    <ActivityIndicator color="white" />
-                </View>
-            ) : requests.length === 0 ? (
-                <View className="flex-1 justify-center items-center">
-                    <Text className="text-white text-base">No friend requests yet.</Text>
-                </View>
-            ) : (
-                <FlatList
-                    data={requests}
-                    keyExtractor={(item) => item.id.toString()}
-                    renderItem={({ item }) => (
-                        <TouchableOpacity
-                          onPress={() => openModal(item)} // sender is the profile of uid1
-                          className="bg-zinc-800 p-4 rounded-xl mb-3"
-                        >
-                            <View className="w-14 h-14 rounded-full items-center justify-center">
-                                {/* Glow layer */}
-                                <View className="absolute w-14 h-14 rounded-full bg-blue-400/30 opacity-60 blur-md" />
-                                <View className="absolute w-12 h-12 rounded-full bg-blue-400/40 blur-sm" />
-                                <Image
-                                    source={{ uri: item.sender.avatar_url }}
-                                    className="w-12 h-12 rounded-full border-2 border-white"
-                                />
-                            </View>
-                            <View className="ml-3 flex-1">
-                                <Text className="text-white text-lg font-semibold">{item.sender.name}</Text>
-                                {item.sender.username && (
-                                    <Text className="text-white/60 text-sm">@{item.sender.username}</Text>
-                                )}
-                            </View>
-                        </TouchableOpacity>
-                      )}
-                />
-            )}
-        </View>
+            <FriendRequestsList onSelect={openModal} />
 
-        <FriendRequestModal
-            visible={modalVisible}
-            request={selectedRequest}
-            onClose={closeModal}
-            onAccept={() => {
-                if (!selectedRequest) return;
-                handleFriendRequest("accept", selectedRequest!.id, selectedRequest!.uid1, selectedRequest!.uid2);
-                closeModal();
-            }}
-            onReject={() => {
-                if (!selectedRequest) return;
-                handleFriendRequest("reject", selectedRequest!.id, selectedRequest!.uid1, selectedRequest!.uid2);
-                closeModal();
-            }}
-            scaleAnim={scaleAnim}
-        />
+            <FriendRequestModal
+                visible={modalVisible}
+                request={selectedRequest}
+                onClose={closeModal}
+                onAccept={() => {
+                    if (!selectedRequest) return;
+                    handleFriendRequest("accept", selectedRequest!.id, selectedRequest!.uid1, selectedRequest!.uid2);
+                    closeModal();
+                }}
+                onReject={() => {
+                    if (!selectedRequest) return;
+                    handleFriendRequest("reject", selectedRequest!.id, selectedRequest!.uid1, selectedRequest!.uid2);
+                    closeModal();
+                }}
+                scaleAnim={scaleAnim}
+            />
+        </View>
     </>
     )
